fix lint

+++ synchronizer/l1_check_block/check_l1block_test.go
@@ -73,7 +73,7 @@ func TestCheckL1BlockHashSafePointIsInFuture(t *testing.T) {
 	data := newTestData(t)
 
 	data.mockState.EXPECT().GetFirstUncheckedBlock(data.ctx, uint64(0), nil).Return(data.stateBlock, nil)
-	data.mockBlockNumberFetch.EXPECT().GetSafeBlockNumber(data.ctx, data.mockL1Client).Return(uint64(data.stateBlock.BlockNumber-1), nil)
+	data.mockBlockNumberFetch.EXPECT().GetSafeBlockNumber(data.ctx, data.mockL1Client).Return(data.stateBlock.BlockNumber-1, nil)
 
 	res := data.sut.Step(data.ctx)
 	require.NoError(t, res)
@@ -83,7 +83,7 @@ func TestCheckL1BlockHashL1ClientReturnsANil(t *testing.T) {
 	data := newTestData(t)
 
 	data.mockState.EXPECT().GetFirstUncheckedBlock(data.ctx, uint64(0), nil).Return(data.stateBlock, nil)
-	data.mockBlockNumberFetch.EXPECT().GetSafeBlockNumber(data.ctx, data.mockL1Client).Return(uint64(data.stateBlock.BlockNumber+10), nil)
+	data.mockBlockNumberFetch.EXPECT().GetSafeBlockNumber(data.ctx, data.mockL1Client).Return(data.stateBlock.BlockNumber+10, nil)
 	data.mockL1Client.EXPECT().HeaderByNumber(data.ctx, big.NewInt(int64(data.stateBlock.BlockNumber))).Return(nil, nil)
 	res := data.sut.Step(data.ctx)
 	require.Error(t, res)
@@ -94,7 +94,7 @@ func TestCheckL1BlockHashMatchHashUpdateCheckMarkOnDB(t *testing.T) {
 	data := newTestData(t)
 
 	data.mockState.EXPECT().GetFirstUncheckedBlock(data.ctx, uint64(0), nil).Return(data.stateBlock, nil)
-	data.mockBlockNumberFetch.EXPECT().GetSafeBlockNumber(data.ctx, data.mockL1Client).Return(uint64(data.stateBlock.BlockNumber), nil)
+	data.mockBlockNumberFetch.EXPECT().GetSafeBlockNumber(data.ctx, data.mockL1Client).Return(data.stateBlock.BlockNumber, nil)
 	l1Block := &types.Header{
 		Number: big.NewInt(100),
 	}
@@ -112,7 +112,7 @@ func TestCheckL1BlockHashMismatch(t *testing.T) {
 
 	data.mockState.EXPECT().GetFirstUncheckedBlock(data.ctx, uint64(0), nil).Return(data.stateBlock, nil)
 	data.stateBlock.BlockHash = common.HexToHash("0x1234") // Wrong hash to trigger a mismatch
-	data.mockBlockNumberFetch.EXPECT().GetSafeBlockNumber(data.ctx, data.mockL1Client).Return(uint64(data.stateBlock.BlockNumber), nil)
+	data.mockBlockNumberFetch.EXPECT().GetSafeBlockNumber(data.ctx, data.mockL1Client).Return(data.stateBlock.BlockNumber, nil)
 	l1Block := &types.Header{
 		Number: big.NewInt(100),
 	}
@@ -122,5 +122,5 @@ func TestCheckL1BlockHashMismatch(t *testing.T) {
 	require.Error(t, res)
 	resErr, ok := res.(*commonsync.ReorgError)
 	require.True(t, ok)
-	require.Equal(t, uint64(data.stateBlock.BlockNumber), resErr.BlockNumber)
+	require.Equal(t, data.stateBlock.BlockNumber, resErr.BlockNumber)
 }
